Disable capitalized-comments 🎉
This rule has so far only caused pain everywhere - many a times you
want to comment out a line of code to try something else. But this
rule then fixes the first keyword on that line to be capitalised
which, when uncommented, would cause a syntax error. Nobody wants
that!

Fixes #14